+++ modules/Shmuelie.PackageManagement/PackageProviders.ps1
@@ -8,6 +8,10 @@ function Get-PackageProvider {
             Get-NpmPackageProvider
             continue
         }
+        if ($name -eq 'Pip') {
+            Get-PipPackageProvider
+            continue
+        }
         [pscustomobject]@{
             Name             = $name
             Platforms        = if ($name -in 'WinGet', 'AppInstaller') { @('Windows') } else { @('Windows', 'Linux', 'MacOS') }